Issue #44: Rename pattern_discovery package to just discovery
