+++ src/main/java/org/wmn4j/mir/discovery/PatternAlgorithms.java
@@ -2,7 +2,7 @@
  * Copyright 2018 Otso Björklund.
  * Distributed under the MIT license (see LICENSE.txt or https://opensource.org/licenses/MIT).
  */
-package org.wmn4j.mir.pattern_discovery;
+package org.wmn4j.mir.discovery;
 
 import java.util.ArrayList;
 import java.util.HashMap;
