+++ src/main/java/org/wmn4j/mir/discovery/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains algorithms for pattern discovery in music notation.
  */
-package org.wmn4j.mir.pattern_discovery;
+package org.wmn4j.mir.discovery;
